postsproject/requirements.txt
asgiref==3.8.1
black==24.4.2
click==8.1.7
colorama==0.4.6
Django==4.2
django-filter==24.2
djangorestframework==3.15.2
drf-yasg==1.21.7
gunicorn==22.0.0
inflection==0.5.1
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
psycopg2==2.9.9
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
